requirements.txt
streamlit
langchain-groq
langgraph
langchain_core
python-dotenv
certifi
httpx
pydantic
fastapi
uvicorn